pyproject.toml
[tool.poetry]
name = "archie-ai-agent"
version = "0.1.0"
description = ""
authors = ["Nikolaj Gordienko <nicolas.gordienko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.1.1"
openai-agents = "^0.2.2"
jinja2 = "^3.1.6"
numpy = "<2.0.0"
sounddevice = "^0.5.2"
simpleaudio = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
